revert: remove chat @ mention picker
The mention picker added to the chat composer did not trigger when `@`
was typed — the popup never showed up in practice. Removing the feature
entirely rather than chasing a debug cycle on a feature that wasn't
providing value.

Deletes:
- packages/chat/src/use-mention-picker.ts
- packages/chat/src/mention-picker.tsx
- packages/chat/src/use-file-attachments.ts

Reverts the mention wiring in chat-input, chat-panel, chat-input-parts,
the package index, and chat-tab.tsx back to their pre-mention state.
